Rename comparators package name
