+++ src/main/java/org/aston/comparators/UserComparator.java
@@ -1,4 +1,4 @@
-package org.aston.programcomparators;
+package org.aston.comparators;
 
 import org.aston.model.User;
 import java.util.Comparator;
